+++ app/minime/urls.py
@@ -8,5 +8,5 @@ app_name = 'minime'
 urlpatterns = [
     url(r'^$', views.index, name='home'),
     url(r'^r/(?P<hash>\w{6})$', views.redirect, name='redirect'),
-    url(r'^shorten/$', views.shorten_url, name='shorten'),
+    url(r'^shorten/$', views.CreateShortUrl.as_view(), name='shorten'),
 ]