+++ HISTORY.md
@@ -3,6 +3,7 @@
 
 ### Next
 
+- Feature: Added Django Rest Framework.
 - Added test coverage and coveralls.
 - Added redis cache.
 - Added ability to run app on Heroku.